refactor(sight): remove AK/SK-based SLS direct upload, keep Logtail file export
- Delete sls.rs (SlsUploader) which used AK/SK env vars to upload
  directly to SLS via aliyun-log-rust-sdk
- Move events_to_flat_records() to logtail.rs so Logtail export
  path continues to work unchanged
- Remove SLS config fields from AgentsightConfig (sls_endpoint,
  sls_access_key_id/secret, sls_project, sls_logstore) and
  their setters/tests
- Remove SLS CLI args from trace subcommand
- Remove aliyun-log-rust-sdk, aliyun-log-sdk-protobuf, tokio
  dependencies from Cargo.toml
- Make Logtail and SQLite mutually exclusive: when SLS_LOGTAIL_FILE
  is set, only Logtail file export is used (skips local SQLite/JSONL)
