+++ src/agentsight/src/genai/mod.rs
@@ -8,7 +8,6 @@ pub mod builder;
 pub mod exporter;
 pub mod storage;
 pub mod instance_id;
-pub mod sls;
 pub mod logtail;
 
 pub use semantic::{
@@ -20,7 +19,6 @@ pub use semantic::{
 pub use exporter::GenAIExporter;
 pub use builder::GenAIBuilder;
 pub use storage::{GenAIStore, GenAIStoreStats};
-pub use sls::SlsUploader;
 pub use logtail::LogtailExporter;
 
 // Blanket implementation: Arc<T> implements GenAIExporter if T does.
